LambdaOpenData/requirements.txt
requests
boto3
opentelemetry-distro